- fixed several compile wornings on BSD   Many thanks to Mark R for providing access to a BSD machine
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@5648 689a6050-402a-0410-94f2-e92a70836424

+++ modules/nat_traversal/nat_traversal.c
@@ -1654,8 +1654,8 @@ save_keepalive_state(void)
             fprintf(f, "%s %.*s %ld %ld\n",
                     contact->uri,
                     contact->socket->sock_str.len, contact->socket->sock_str.s,
-                    contact->registration_expire,
-                    contact->subscription_expire);
+                    (long)contact->registration_expire,
+                    (long)contact->subscription_expire);
             contact = contact->next;
         }
     }
